FastAPI_version/day1-ip-api/pyproject.toml
[tool.poetry]
name = "fastapi-version"
version = "0.1.0"
description = "FastAPI CRUD API project"
authors = ["B <email@hotmail.com>"]
readme = "README.md"

packages = [
    { include = "fastapi_version" }
]


[tool.poetry.dependencies]
python = ">=3.13"
fastapi = ">=0.123.0,<0.124.0"
uvicorn = ">=0.38.0,<0.39.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
httpx = "^0.28.1"
pytest = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = ">=9.0.1,<10.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

